LF-3622: mtd: nand: gpmi: fix the unintentional integer overflow issue
cast the variable from int to loff_t to avoid potential overflow issue,
reported by coverity. CID 17332.

+++ drivers/mtd/nand/raw/gpmi-nand/gpmi-nand.c
@@ -2249,7 +2249,7 @@ static int mx23_boot_init(struct gpmi_nand_data  *this)
 		 */
 		chipnr = block >> (chip->chip_shift - chip->phys_erase_shift);
 		page = block << (chip->phys_erase_shift - chip->page_shift);
-		byte = block <<  chip->phys_erase_shift;
+		byte = (loff_t)block <<  chip->phys_erase_shift;
 
 		/* Send the command to read the conventional block mark. */
 		nand_select_target(chip, chipnr);
